Fix infamous window resizing bug
This commit fixes a bug on Mac where the game loop hangs during window resize. This could cause the player to be kicked from the server due to inactivity (no packets sent in over a second) and it also causes the UI to freeze and stretch. The network issue could be solved with ping packets, but I wanted to solve the underlying problem instead, so here we go:

To fix these issues the main application thread was split in two:
1. The main thread only runs code that has to be run on the main thread (mostly window management code).
2. The game thread runs the game loop (which was previously run by the main thread) but can ask the main thread to perform tasks if needed. It then waits for the main thread to run it, except in some cases where the call has been wrapped in a Future in order to not block the game loop.

The reason for all of this back-and-forth is that most functions in GLFW (window manager) have to be run on the main thread and all functions in OpenGL have to be run on only one thread (the game thread in this case).

In addition to all of this Apple has also decided that one has to lock the OpenGL context before using it, and if you don't you will get various memory corruption issues in the native code and the game will likely crash. Luckily this was quite easy to fix.

At the end of all of this, it actually works pretty well!